Wave A: lint our register headers + tests/mocks; strip 2.3k redundant casts
Drop ra8d2_*_regs and tests/mocks from .clang-tidy HeaderFilterRegex --
both are first-party code and should be linted. The previous exclusion
silenced ~2300 redundant-casting findings from typed-enum constants,
e.g. (uint16_t)k_some_enum where the enum is already `: uint16_t`.

clang-tidy --fix removed all of them across 244 files. Builds and the
full host-test suite are unchanged; the casts were genuinely no-ops.
Cleanup of the auto-fix collateral:

- The fixer used C++ `static_cast<int>` syntax in C files when widening
  bool to int, then redundantly wrapped expressions like `if (X)` as
  `(int)(X)`. Removed the wrappers entirely (C accepts any scalar in
  conditional context) and dropped the redundant parens (which v22's
  readability-redundant-parentheses correctly flagged).
- `readability-implicit-bool-conversion` flagged idiomatic C `return x
  != y;` because relational operators in C produce `int` per the C
  spec, and its v18 auto-fixer wrapped them in `(x != y) != 0` --
  identical noise. Reverted the 99 spurious wrappers and disabled the
  check with a comment explaining the C/C++ semantic mismatch.

Three computed-cast sites in ra_rsip.c / ra8d2_mipi_csi_regs.h needed
NOLINT for clang-analyzer-optin.core.EnumCastOutOfRange (NOLINT count
42 -> 57) -- the analyzer can't see that a base + stride lands on a
HUM-defined register, only that the literal value isn't an enumerator.

Per-line rationale comments now sit above each disabled check in
.clang-tidy so the next reader knows why each one is off.

Tidy now passes on both clang-tidy 18 (Ubuntu/CI/docker) and 22
(Homebrew/macOS); 86/86 host tests pass; blink + blink_hal sizes
unchanged.

+++ blink_hal/main.c
@@ -166,7 +166,7 @@ int32_t main(void)
 {
   /* Driver init runs with PRIMASK set (SystemInit masks IRQs at boot
    * so partial state isn't accidentally driven by an interrupt). */
-  if (ra_time_init((uint32_t)k_blink_cpu_hz_at_reset) != k_ra_ok) {
+  if (ra_time_init(k_blink_cpu_hz_at_reset) != k_ra_ok) {
     while (1) {
       __asm__ volatile("wfi");
     }
@@ -187,7 +187,7 @@ int32_t main(void)
     if (blink_pins_toggle_all() != k_ra_ok) {
       break;
     }
-    ra_delay_ms((uint32_t)k_blink_half_period_ms);
+    ra_delay_ms(k_blink_half_period_ms);
   }
 
   while (1) {
