+++ libs/ra_hal/src/adc.c
@@ -55,8 +55,7 @@ static ra_adc_state_t s_adc_state;
  */
 static uint32_t internal_ressel(ra_adc_resolution_t resolution)
 {
-  return ((uint32_t)resolution << (uint32_t)k_ra_adchcr_bit_ressel0) &
-         (uint32_t)k_ra_adchcr_mask_ressel;
+  return ((uint32_t)resolution << (uint32_t)k_ra_adchcr_bit_ressel0) & k_ra_adchcr_mask_ressel;
 }
 
 /**
@@ -65,12 +64,12 @@ static uint32_t internal_ressel(ra_adc_resolution_t resolution)
 static void internal_apply_resolution_all(ra_adc_resolution_t resolution)
 {
   const uint32_t field = internal_ressel(resolution);
-  for (uint8_t ch = 0U; ch < (uint8_t)k_ra_adc_b_max_channels; ++ch) {
+  for (uint8_t ch = 0U; ch < k_ra_adc_b_max_channels; ++ch) {
     volatile uint32_t* reg = ra_adc_b_adchcr(ch);
     if (reg == nullptr) {
       continue;
     }
-    *reg = ((*reg & ~(uint32_t)k_ra_adchcr_mask_ressel) | field);
+    *reg = ((*reg & ~k_ra_adchcr_mask_ressel) | field);
   }
 }
 
@@ -81,9 +80,9 @@ ra_err_t ra_adc_init(void)
   RA_RETURN_ON_ERROR(mst_err, s_tag, "adc_init: mstp enable"); /* GCOVR_EXCL_BR_LINE */
 
   /* HUM Ch 53 "16-bit A/D Converter (ADC16H)" p 3308 */
-  *ra_adc_b_adclkenr() = (uint32_t)k_ra_adclkenr_mask_clken;
+  *ra_adc_b_adclkenr() = k_ra_adclkenr_mask_clken;
   *ra_adc_b_adclkcr()  = 0U;
-  *ra_adc_b_admdr()    = (uint32_t)k_ra_adc_admdr_default;
+  *ra_adc_b_admdr()    = k_ra_adc_admdr_default;
   *ra_adc_b_adintcr()  = 0U;
 
   /* Default to 14-bit on every channel config slot. */
@@ -109,11 +108,11 @@ ra_err_t ra_adc_read_channel(uint8_t channel, uint16_t* out_raw)
    * completion. In the sim build the mmap backing stays asserted
    * which means the poll loop below exits on the first iteration
    * with whatever value the test pre-seeded into ADDR[channel]. */
-  *chcr = (*chcr | (uint32_t)k_ra_adchcr_mask_cven);
+  *chcr = (*chcr | k_ra_adchcr_mask_cven);
 
-  for (uint32_t i = 0U; i < (uint32_t)k_ra_adc_poll_limit; ++i) {
-    if ((*chcr & (uint32_t)k_ra_adchcr_mask_cven) == 0U) {
-      *out_raw = (uint16_t)(*addr & (uint32_t)k_ra_adc_result_mask);
+  for (uint32_t i = 0U; i < k_ra_adc_poll_limit; ++i) {
+    if ((*chcr & k_ra_adchcr_mask_cven) == 0U) {
+      *out_raw = (uint16_t)(*addr & k_ra_adc_result_mask);
       return k_ra_ok;
     }
   }
@@ -130,15 +129,15 @@ ra_err_t ra_adc_init_configured(const ra_adc_cfg_t* cfg)
   const ra_err_t mst_err = ra_mstp_enable(k_ra_mstp_adc16h);
   RA_RETURN_ON_ERROR(mst_err, s_tag, "adc_init_cfg: mstp enable");
 
-  *ra_adc_b_adclkenr() = (uint32_t)k_ra_adclkenr_mask_clken;
+  *ra_adc_b_adclkenr() = k_ra_adclkenr_mask_clken;
   *ra_adc_b_adclkcr()  = 0U;
 
-  uint32_t mdr = (uint32_t)k_ra_adc_admdr_default;
+  uint32_t mdr = k_ra_adc_admdr_default;
   if (cfg->scan_mode) {
-    mdr |= (uint32_t)k_ra_adc_admdr_scan_bit;
+    mdr |= k_ra_adc_admdr_scan_bit;
   }
   if ((cfg->trigger == k_ra_adc_trig_external) || (cfg->trigger == k_ra_adc_trig_elc)) {
-    mdr |= (uint32_t)k_ra_adc_admdr_trigext_bit;
+    mdr |= k_ra_adc_admdr_trigext_bit;
   }
   *ra_adc_b_admdr()   = mdr;
   *ra_adc_b_adintcr() = 0U;
@@ -163,7 +162,7 @@ ra_err_t ra_adc_deinit(void)
 
 ra_err_t ra_adc_set_resolution(ra_adc_resolution_t resolution)
 {
-  if ((uint8_t)resolution > (uint8_t)k_ra_adc_res_14bit) {
+  if ((uint8_t)resolution > k_ra_adc_res_14bit) {
     return k_ra_err_invalid_arg;
   }
   internal_apply_resolution_all(resolution);
@@ -175,11 +174,11 @@ ra_err_t ra_adc_get_status(uint16_t* out_mask)
   RA_CHECK_NULL_PTR(out_mask, s_tag, "out_mask must not be nullptr");
   const uint32_t mdr = *ra_adc_b_admdr();
   uint16_t       out = 0U;
-  if ((mdr & (uint32_t)k_ra_admdr_mask_adbusy) != 0U) {
-    out |= (uint16_t)k_ra_adc_status_busy;
+  if ((mdr & k_ra_admdr_mask_adbusy) != 0U) {
+    out |= k_ra_adc_status_busy;
   }
   if (*ra_adc_b_adintcr() != 0U) {
-    out |= (uint16_t)k_ra_adc_status_ie;
+    out |= k_ra_adc_status_ie;
   }
   *out_mask = out;
   return k_ra_ok;
@@ -190,7 +189,7 @@ ra_err_t ra_adc_clear_status(void)
   /* Clearing "busy" means stopping any in-flight conversion. Mask the
    * top bit off ADMDR. (ADBUSY is strictly RO on real silicon, but the
    * driver models it for tests / future ADMDR write-to-clear bits.) */
-  *ra_adc_b_admdr() = (*ra_adc_b_admdr() & ~(uint32_t)k_ra_admdr_mask_adbusy);
+  *ra_adc_b_admdr() = (*ra_adc_b_admdr() & ~k_ra_admdr_mask_adbusy);
   return k_ra_ok;
 }
 
@@ -218,7 +217,7 @@ void ra_adc_dispatch_cnv_end(uint8_t channel)
   if (addr == nullptr) {
     return;
   }
-  const uint16_t             result = (uint16_t)(*addr & (uint32_t)k_ra_adc_result_mask);
+  const uint16_t             result = (uint16_t)(*addr & k_ra_adc_result_mask);
   const ra_adc_complete_fn_t fn     = s_adc_state.fn;
   void* const                ctx    = s_adc_state.ctx;
   if (fn != nullptr) {
